@@ -4,12 +4,12 @@
 
 	<name>money-utility</name>
 	<description>Implementation of the Money class as described by Martin Fowler</description>
-	<url>https://github.com/leandroaparecido/money</url>
+	<url>https://github.com/lehphyro/money</url>
 
 	<parent>
 		<groupId>org.sonatype.oss</groupId>
 		<artifactId>oss-parent</artifactId>
-		<version>7</version>
+		<version>9</version>
 	</parent>
 
 	<groupId>com.github.lehphyro.money</groupId>
@@ -20,20 +20,20 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
-			<version>12.0</version>
+			<version>27.1-jre</version>
 		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
-			<version>4.8.2</version>
+			<version>4.12</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
 
 	<scm>
-		<connection>scm:git:git@github.com:leandroaparecido/money.git</connection>
-		<developerConnection>scm:git:git@github.com:leandroaparecido/money.git</developerConnection>
-		<url>git@github.com:leandroaparecido/money.git</url>
+		<connection>scm:git:git@github.com:lehphyro/money.git</connection>
+		<developerConnection>scm:git:git@github.com:lehphyro/money.git</developerConnection>
+		<url>git@github.com:lehphyro/money.git</url>
 	</scm>
 
 	<build>
@@ -41,50 +41,12 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.5.1</version>
+				<version>3.8.0</version>
 				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
+					<source>1.8</source>
+					<target>1.8</target>
 				</configuration>
 			</plugin>
 		</plugins>
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>org.eclipse.m2e</groupId>
-					<artifactId>lifecycle-mapping</artifactId>
-					<version>1.0.0</version>
-					<configuration>
-						<lifecycleMappingMetadata>
-							<pluginExecutions>
-								<pluginExecution>
-									<pluginExecutionFilter>
-										<groupId>org.apache.maven.plugins</groupId>
-										<artifactId>maven-enforcer-plugin</artifactId>
-										<versionRange>[1.0.0,)</versionRange>
-										<goals>
-											<goal>enforce</goal>
-										</goals>
-									</pluginExecutionFilter>
-									<action>
-										<ignore />
-									</action>
-								</pluginExecution>
-							</pluginExecutions>
-						</lifecycleMappingMetadata>
-					</configuration>
-				</plugin>
-			</plugins>
-		</pluginManagement>
 	</build>
-
-	<developers>
-		<developer>
-			<id>leandro</id>
-			<email>lehphyro@gmail.com</email>
-			<roles>
-				<role>leader</role>
-			</roles>
-		</developer>
-	</developers>
 </project>
